AddCommentInWord/requirements.txt
pandas
openpyxl
python-docx
PyPDF2
reportlab